[Robot] Use last valid path until a new valid path is generated
Because pathfinder produces the most optimal path, this means that the robot is as close to the area that the pathfinder believes is impassable as possible. Due to this, the robot sometimes wanders into a bad area but is still able to keep going.

The new expected behavior is that the robot will continue on the last valid path that it produced. This has the added benefit of allowing the robot to continue its path, even if the Raspberry Pi dies during the run.