[Robot] Add cone scoring commands
These likely need to be tuned as the speed of the intake has been changed, as have the positions for the arm.

Scoring cones high does not work in the sim in teleop or in auto. I would guess that this is because it is trying to go to an illegal pose.